w3c-reports/12092014/mac/runned-tests/workers/interfaces/WorkerGlobalScope/onerror/not-handled-expected.txt


FAIL onerror, "not handled" assert_equals: typeof e.column expected "number" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/workers/interfaces/WorkerGlobalScope/onerror/not-handled.html:24:18
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
